refactor:tests questions and answer with attachments

+++ src/infra/database/prisma/repositories/prisma-questions-attachments-repository.ts
@@ -5,24 +5,55 @@ import { PrismaService } from "../prisma.service";
 import { PrismaQuestionAttachmentsMapper } from "../mappers/prisma-question-attachment-mapper";
 
 @Injectable()
-export class PrismaQuestionsAttachmentsRepository implements QuestionAttachmentRepository{
-    constructor(private readonly prisma: PrismaService) {}
+export class PrismaQuestionsAttachmentsRepository implements QuestionAttachmentRepository {
+  constructor(private readonly prisma: PrismaService) {}
 
-    async findManyByQuestionId(questionId: string): Promise<QuestionAttachment[]> {
-        const questionAttachments = await this.prisma.attachment.findMany({
-            where: {
-                questionId
-            }
-        }) 
+  async findManyByQuestionId(
+    questionId: string,
+  ): Promise<QuestionAttachment[]> {
+    const questionAttachments = await this.prisma.attachment.findMany({
+      where: {
+        questionId,
+      },
+    });
 
-        return questionAttachments.map(PrismaQuestionAttachmentsMapper.toDomain)
+    return questionAttachments.map(PrismaQuestionAttachmentsMapper.toDomain);
+  }
+
+  async deleteManyByQuestionId(questionId: string): Promise<void> {
+    await this.prisma.attachment.deleteMany({
+      where: {
+        questionId,
+      },
+    });
+  }
+
+  async createMany(attachments: QuestionAttachment[]): Promise<void> {
+    if (attachments.length === 0) {
+      return;
     }
 
-    async deleteManyByQuestionId(questionId: string): Promise<void> {
-        await this.prisma.attachment.deleteMany({
-            where: {
-                questionId
-            }
-        })
+    const data =
+      PrismaQuestionAttachmentsMapper.toPrismaUpdateMany(attachments);
+
+    await this.prisma.attachment.updateMany(data);
+  }
+
+  async deleteMany(attachments: QuestionAttachment[]): Promise<void> {
+    if (attachments.length === 0) {
+      return;
     }
+
+    const attachmentIds = attachments.map((attachment) => {
+      return attachment.id.toString();
+    });
+
+    await this.prisma.attachment.deleteMany({
+        where: {
+            id: {
+                in: attachmentIds
+            }
+        }
+    })
+  }
 }
